requirements.txt
yamllint==1.17.0
jsonschema==3.0.2
PyYAML==5.1.2
